requirements.txt
absl-py @ file:///croot/absl-py_1686852429912/work
aiohttp @ file:///croot/aiohttp_1701112538292/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
anyio==4.2.0
array-record==0.5.0
astunparse==1.6.3
async-timeout @ file:///opt/conda/conda-bld/async-timeout_1664876359750/work
attrs @ file:///croot/attrs_1695717823297/work
blinker @ file:///croot/blinker_1696539051175/work
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///croot/click_1698129812380/work
cryptography @ file:///croot/cryptography_1694444244250/work
dill @ file:///croot/dill_1692271232022/work
dm-tree==0.1.8
etils==1.6.0
exceptiongroup==1.2.0
flatbuffers==23.5.26
frozenlist @ file:///croot/frozenlist_1698702560391/work
fsspec==2023.12.2
future @ file:///croot/future_1677599870788/work
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
google-auth @ file:///croot/google-auth_1694152708165/work
google-auth-oauthlib @ file:///opt/conda/conda-bld/google-auth-oauthlib_1660687784486/work
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
googleapis-common-protos @ file:///opt/conda/conda-bld/googleapis-common-protos-feedstock_1660256159732/work
grpcio @ file:///croot/grpc-suite_1681912592597/work
h11==0.14.0
h5py @ file:///croot/h5py_1691589708553/work
httpcore==1.0.2
httpx==0.25.2
idna @ file:///croot/idna_1666125576474/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1701632192416/work
importlib-resources==6.1.1
jax==0.4.23
joblib @ file:///croot/joblib_1685113087166/work
keras==2.15.0
Keras-Preprocessing @ file:///home/conda/feedstock_root/build_artifacts/keras-preprocessing_1610713559828/work
libclang==16.0.6
Markdown @ file:///croot/markdown_1671541909495/work
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
mkl-service==2.4.0
ml-dtypes==0.2.0
multidict @ file:///croot/multidict_1701096859099/work
numexpr @ file:///croot/numexpr_1696515281613/work
numpy==1.22.4
nvidia-cublas-cu12==12.2.5.6
nvidia-cuda-cupti-cu12==12.2.142
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu12==12.2.140
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cudnn-cu12==8.9.4.25
nvidia-cufft-cu12==11.0.8.103
nvidia-curand-cu12==10.3.3.141
nvidia-cusolver-cu12==11.5.2.141
nvidia-cusparse-cu12==12.1.2.141
nvidia-nccl-cu12==2.16.5
nvidia-nvjitlink-cu12==12.2.140
oauthlib @ file:///croot/oauthlib_1679489621486/work
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging @ file:///croot/packaging_1693575174725/work
pandas @ file:///croot/pandas_1702317985682/work/dist/pandas-2.1.4-cp310-cp310-linux_x86_64.whl#sha256=dcdb9c0337e13fd44ae36b0cc0437403549fd0bc4ab98d7550aa3564748c6736
promise @ file:///home/builder/ci_310/promise_1640792568485/work
protobuf==3.20.3
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT @ file:///opt/conda/conda-bld/pyjwt_1657544592787/work
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-telegram-bot==20.7
pytz @ file:///croot/pytz_1695131579487/work
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
requests @ file:///croot/requests_1690400202158/work
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-learn @ file:///croot/scikit-learn_1690978916802/work
scipy @ file:///croot/scipy_1701295040508/work/dist/scipy-1.11.4-cp310-cp310-linux_x86_64.whl#sha256=98e4e3a1ef49d0edb98b5eb8384c7c5204a5378bf8b56aa1c6a5b03208b9d91a
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
tensorboard==2.15.1
tensorboard-data-server @ file:///croot/tensorboard-data-server_1681498183723/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl
tensorboard-plugin-wit @ file:///home/builder/tkoch/workspace/tensorflow/tensorboard-plugin-wit_1658918494740/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow==2.15.0.post1
tensorflow-datasets==4.9.4
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
tensorflow-metadata==1.14.0
termcolor @ file:///croot/termcolor_1668084651543/work
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
toml==0.10.2
tqdm @ file:///croot/tqdm_1679561862951/work
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///croot/urllib3_1698257533958/work
Werkzeug @ file:///croot/werkzeug_1679489717957/work
wrapt @ file:///tmp/abs_c335821b-6e43-4504-9816-b1a52d3d3e1eel6uae8l/croots/recipe/wrapt_1657814400492/work
yarl @ file:///croot/yarl_1701105127787/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
